Illustrate fix for #4481 in terms of an 0.9 patch
Release 0.9 does not include the more comprehensive textual
coercion system introduced in 1.0 in #2992, however we present
here a blunt patch to disallow text coercion in order_by()
and group_by() completely.  This breaks backwards compatibility
with applications that rely on this feature, including that it
also removes the feature in 1.0 where strings that match existing
column names are always accepted and matched to the column without
any textual coercion.

This patch may be used by downstream distros who wish to bluntly
disallow the order_by()/group_by() coercion completely, without
regard for applications that may make use of this relatively obscure
feature.

Change-Id: I3363b21428f1ad8797394b63197375a2e56a0bd7
References: #4481